The export's ffmpeg passes, in one place, with the new one beside them
exportVideoClips moves out of VideoProcessingService into video-export-passes,
a factory over the three collaborators it needs, and the Definition Card
compositing pass lands next to it. The rest of that service is OBS,
transcription and DaVinci Resolve, which have nothing to do with an export;
now the export pipeline reads end to end in one file — and fits under the
file-token limit, which it no longer did.

compositeOverlaysOntoExport writes beside the export and moves the finished
file over it, because ffmpeg cannot read and write one file at once. It is a
pass of its own, on the flattened timeline AFTER the concat, rather than
something fused into the per-Clip crop filter Clip Zoom uses: a future
Transform has to interpolate across a Clip boundary, which a pre-concat
per-input filter cannot express, so the stage is placed now for it to land in.
