+++ apps/local/app/services/video-processing-service.ts
@@ -10,12 +10,8 @@ import { FFmpegCommandsService } from "./ffmpeg-commands";
 import { findSilenceInVideo } from "./silence-detection";
 import { transcribeFootage } from "./footage-transcription";
 import { VideoEditorLoggerService } from "./video-editor-logger-service";
-import { makeFfmpegLogger } from "./ffmpeg-video-logger";
+import { makeVideoExportPasses } from "./video-export-passes";
 import type { SilenceLength } from "@/silence-detection-constants";
-import {
-  VIDEO_FORMAT_DIMENSIONS,
-  type VideoFormat,
-} from "@/features/videos/video-format";
 
 export type PauseType = "none" | "long";
 
@@ -133,98 +129,8 @@ export class VideoProcessingService extends Effect.Service<VideoProcessingServic
         }
       );
 
-      const exportVideoClips = Effect.fn("exportVideoClips")(function* (opts: {
-        videoId: string;
-        format: VideoFormat;
-        clips: {
-          inputVideo: string;
-          startTime: number;
-          duration: number;
-          pauseType: PauseType;
-          zoomType: string;
-        }[];
-        shortsDirectoryOutputName: string | undefined;
-        onStageChange?: (
-          stage: "concatenating-clips" | "normalizing-audio"
-        ) => void;
-        /**
-         * Real per-phase progress from the underlying ffmpeg processes.
-         * `percent` is an integer 0–99 that resets when the stage changes;
-         * 100 is signalled by completion, not by this callback.
-         */
-        onProgress?: (info: {
-          stage: "concatenating-clips" | "normalizing-audio";
-          percent: number;
-        }) => void;
-      }) {
-        const FINISHED_VIDEOS_DIRECTORY = yield* Config.string(
-          "FINISHED_VIDEOS_DIRECTORY"
-        );
-
-        // Every ffmpeg invocation for this export is teed into
-        // `.data/logs/{videoId}.log` (fetch its path via VideoEditorLoggerService
-        // or GET /api/videos/:videoId/log-path) — the "cli-output" event, on
-        // both success and failure, so a rare hang or corrupt export has a
-        // durable artifact to diagnose from instead of a swallowed exit code.
-        const logCliOutput = (stage: "concat" | "normalize-audio") =>
-          makeFfmpegLogger(videoEditorLogger, opts.videoId, `export:${stage}`);
-
-        // Create concatenated video using native FFmpeg, in the aspect ratio
-        // that matches the video's format (portrait for shorts, landscape
-        // otherwise).
-        opts.onStageChange?.("concatenating-clips");
-        const concatenatedPath =
-          yield* ffmpegCommands.createAndConcatenateVideoClipsSinglePass(
-            opts.clips,
-            VIDEO_FORMAT_DIMENSIONS[opts.format],
-            {
-              onProgress: (percent) =>
-                opts.onProgress?.({ stage: "concatenating-clips", percent }),
-              onLog: logCliOutput("concat"),
-            }
-          );
-
-        // Normalize audio
-        opts.onStageChange?.("normalizing-audio");
-        const normalizedPath = yield* ffmpegCommands.normalizeAudio(
-          concatenatedPath,
-          {
-            onProgress: (percent) =>
-              opts.onProgress?.({ stage: "normalizing-audio", percent }),
-            onLog: logCliOutput("normalize-audio"),
-          }
-        );
-
-        // Move to final location
-        const outputPath = path.join(
-          FINISHED_VIDEOS_DIRECTORY,
-          `${opts.videoId}.mp4`
-        );
-
-        yield* effectFs.makeDirectory(path.dirname(outputPath), {
-          recursive: true,
-        });
-        // Use copy+remove instead of rename to support cross-device moves
-        // (e.g. /tmp on tmpfs → /mnt/d on NTFS via WSL2)
-        yield* effectFs.copyFile(normalizedPath, outputPath);
-
-        // Measure what was actually produced. ffmpeg exiting zero says only
-        // that it stopped without complaining, not that it wrote every frame
-        // it was asked for — so the caller, which knows what the Clips asked
-        // for, is handed the real number and refuses a short file.
-        const durationInSeconds =
-          yield* ffmpegCommands.getVideoDurationInSeconds(outputPath);
-
-        // Clean up intermediate files
-        yield* effectFs
-          .remove(normalizedPath)
-          .pipe(Effect.catchAll(() => Effect.void));
-        yield* effectFs
-          .remove(concatenatedPath)
-          .pipe(Effect.catchAll(() => Effect.void));
-
-        return { outputPath, durationInSeconds };
-      });
+      const { exportVideoClips, compositeOverlaysOntoExport } =
+        makeVideoExportPasses({ ffmpegCommands, effectFs, videoEditorLogger });
 
       /**
        * Extract audio from a video clip segment using ffmpeg.
@@ -634,6 +540,7 @@ export class VideoProcessingService extends Effect.Service<VideoProcessingServic
       return {
         getLatestOBSVideoClips,
         exportVideoClips,
+        compositeOverlaysOntoExport,
         /**
          * The container duration of a finished file, in seconds.
          *
